pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proto-v2-engine"
version = "0.0.0"
dependencies = ["pydantic>=2.0", "fastapi>=0.100", "jsonschema>=4.21"]
requires-python = ">=3.9"

[project.optional-dependencies]
test = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "PyYAML>=6.0",
  "alembic>=1.13",
  "asgi-lifespan>=2.1",
  "aiosqlite>=0.20",
  "httpx>=0.27",
  "greenlet>=3.0",
  "anyio>=4.0",
  "sqlalchemy>=2.0",
  "jsonschema>=4.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
exclude = [
  "data",
  "data.*",
  "logs",
  "logs.*",
  "jobs",
  "jobs.*",
]
